data-ingestion/requirements.txt
finnhub-python==2.4.19
numpy==1.26.1
pandas==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.4
